Fixed hazard detection bug for branches
The hazard unit wouldn't wait for branches to resolve their inputs UNLESS the branch was going to be taken. This probably didn't come up sooner because branches are taken so often.
This bug is independent of the branch code refactoring.

An interesting detail while I was fixing this: BranchD was already declared in cpu.v, with a note explaining this exact bug. But it was never wired up to anything. Huh.